fix payment

+++ homesharing_back-end/src/main/java/com/homesharing_backend/service/impl/PaymentServiceImpl.java
@@ -52,7 +52,7 @@ public class PaymentServiceImpl implements PaymentService {
         if (Objects.isNull(paymentPackage)) {
             throw new NotFoundException("khong tim thay payment-packate-id");
         } else {
-            float price = paymentPackage.getPrice() * 100;
+            int price = paymentPackage.getPrice() * 100;
 
             String paymentID = paymentRequest.getPaymentPackageID() + "-"
                     + paymentRequest.getPostID() + "-" + PaymentConfig.getRandomNumber(6);
